requirements/base.txt
django
mysqlclient
django-simpleui
djangorestframework
django-cors-headers

djangorestframework-jwt
django-guardian
Pillow